+++ docs/fr_FR/changelog.md
@@ -1,3 +1,4 @@
 # 20/02/2018
 
+- Ajout d'un bouton pour envoyer la configuration automatiquement au market
 - Corrections de bugs